feat(ui): introduce app templates
