+++ ui/src/modules/app_list/index.tsx
@@ -16,24 +16,23 @@ import {
   Title
 } from '@mantine/core'
 import { IconDots, IconPlus, IconSearch } from '@tabler/icons-react'
-import { getHotkeyHandler, useClipboard, useDisclosure } from '@mantine/hooks'
+import { useClipboard, useDisclosure } from '@mantine/hooks'
 import {
   getListAppApplicationsGetQueryKey,
-  useCreateAppApplicationsPost,
   useDeleteAppApplicationsApplicationIdDelete,
-  useListAppApplicationsGet,
-  useUpdateAppMetaApplicationsApplicationIdPut
+  useListAppApplicationsGet
 } from '@api/linguflow'
 import { ApplicationInfo } from '@api/linguflow.schemas'
-import { useEffect, useMemo, useState } from 'react'
+import { useMemo, useState } from 'react'
 import dayjs from 'dayjs'
 import { useQueryClient } from 'react-query'
-import { Link, useNavigate } from 'react-router-dom'
+import { Link } from 'react-router-dom'
 import { Card, LoadingCard } from '../../components/Card'
 import { Pagination } from '../../components/Pagination'
 
 import { NoResult } from '../../components/NoResult'
 import classes from './index.module.css'
+import { ModifyAppModel } from './ModifyAppModel'
 
 const PAGE_SIZE = 12
 
@@ -87,131 +86,6 @@ export const AppList: React.FC = () => {
         )}
       </Stack>
     </Container>
-  )
-}
-
-interface ModifyAppModelProps {
-  opened: boolean
-  onClose: () => void
-  app?: ApplicationInfo
-}
-
-const ModifyAppModel: React.FC<ModifyAppModelProps> = ({ opened, onClose, app }) => {
-  const queryClient = useQueryClient()
-  const navigate = useNavigate()
-  const [name, setName] = useState(app?.name || '')
-  const [langfusePK, setLangfusePK] = useState(app?.langfuse_public_key || '')
-  const [langfuseSK, setLangfuseSK] = useState(app?.langfuse_secret_key || '')
-  const { mutateAsync: createApp, isLoading: isCreating } = useCreateAppApplicationsPost({
-    mutation: {
-      onSuccess: async (data) => {
-        await queryClient.fetchQuery({ queryKey: getListAppApplicationsGetQueryKey() })
-        onClose()
-        navigate(`/app/${data.id}`)
-      }
-    }
-  })
-  const { mutateAsync: updateApp, isLoading: isUpdating } = useUpdateAppMetaApplicationsApplicationIdPut({
-    mutation: {
-      onSuccess: async () => {
-        await queryClient.fetchQuery({ queryKey: getListAppApplicationsGetQueryKey() })
-        onClose()
-      }
-    }
-  })
-  const handleConfirm = async () => {
-    if (isLoading || !opened || !name) {
-      return
-    }
-
-    if (!app) {
-      await createApp({ data: { name, langfusePublicKey: langfusePK, langfuseSecretKey: langfuseSK } })
-    } else {
-      await updateApp({
-        applicationId: app.id,
-        data: { name, langfusePublicKey: langfusePK, langfuseSecretKey: langfuseSK }
-      })
-    }
-  }
-
-  const isLoading = isCreating || isUpdating
-
-  useEffect(() => {
-    if (!opened) {
-      return
-    }
-    setName(app?.name || '')
-    setLangfusePK(app?.langfuse_public_key || '')
-    setLangfuseSK(app?.langfuse_secret_key || '')
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [opened])
-
-  return (
-    <Modal
-      closeOnClickOutside={!isLoading}
-      closeOnEscape={!isLoading}
-      withCloseButton={!isLoading}
-      opened={opened}
-      onClose={onClose}
-      title={
-        <Text fw="bold">
-          {!app ? (
-            'New Application'
-          ) : (
-            <>
-              Edit <Code fz="md">{app.name}</Code>
-            </>
-          )}
-        </Text>
-      }
-      centered
-      trapFocus={false}
-    >
-      <Stack>
-        <TextInput
-          required
-          autoFocus
-          label="Name"
-          placeholder="Please input the application name"
-          value={name}
-          disabled={isLoading}
-          onChange={(event) => {
-            setName(event.currentTarget.value)
-          }}
-          onKeyDown={getHotkeyHandler([['Enter', handleConfirm]])}
-        />
-
-        <TextInput
-          label="Langfuse Public Key"
-          placeholder="Please input the langfuse public key"
-          value={langfusePK}
-          disabled={isLoading}
-          onChange={(event) => {
-            setLangfusePK(event.currentTarget.value)
-          }}
-        />
-
-        <TextInput
-          label="Langfuse Secret Key"
-          placeholder="Please input the langfuse secret key"
-          value={langfuseSK}
-          disabled={isLoading}
-          onChange={(event) => {
-            setLangfuseSK(event.currentTarget.value)
-          }}
-          onKeyDown={getHotkeyHandler([['Enter', handleConfirm]])}
-        />
-      </Stack>
-
-      <Group mt="xl" justify="end">
-        <Button variant="default" onClick={onClose} disabled={isLoading}>
-          Cancel
-        </Button>
-        <Button color="dark" loading={isLoading} disabled={!name} onClick={handleConfirm}>
-          Confirm
-        </Button>
-      </Group>
-    </Modal>
   )
 }
